+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added developerId, developerName, developerEmail, scmConnection, scmDeveloperConnection, scmUrl fields to `BomMetadata`.
 - Added `licenses` tag to generated `pom.xml` when the user fills licenses info in `BomMetadata`.
 - Added `developers` tag to generated `pom.xml` when the user fills developer info in `BomMetadata`.
+- Added `scm` tag to generated `pom.xml` when the user fills scm info in `BomMetadata`.
 
 ### Modified
 ### Removed